Implement random bits and move public API to main module
