+++ test/prng_test.exs
@@ -1,0 +1,4 @@
+defmodule PRNGTest do
+  use ExUnit.Case, async: true
+  doctest PRNG
+end
