--- test/threefry_test.exs
@@ -1,4 +1,0 @@
-defmodule PRNG.ThreefryTest do
-  use ExUnit.Case, async: true
-  doctest PRNG.Threefry
-end
